Add inherit-raw and inherit-leaky-raw
Introduces new inheritance keys, :inherit-raw and :inherit-leaky-raw,
which pull inherited values from the raw project map. This is useful
when wanting to inherit values that are transformed upon project map
load, like paths, which Leiningen makes absolute when loading a project.

Includes updates to with-all to make to enable this behavior when
creating the :monolith/all profile.