Set initial thread count to avoid running tasks sequentially for 10 seconds (#102)
The default for a manifold fixed thread pool is that it starts with 1 thread, and then every `control-period` adjusts the thread count. The default `control-period` is 10 seconds, so for the first 10 seconds, `each :parallel` will process projects sequentially. This sets `:initial-thread-count` to the desired number of threads to start with the full desired parallelism.